// src/app/models/brewery.model.ts
export interface Brewery {
    id: string;
    name: string;
    brewery_type: string | null;
    address_1: string | null;
    address_2: string | null;
    address_3: string | null;
    city: string;
    state_province: string;
    postal_code: string | null;
    country: string;
    longitude: string | null;
    latitude: string | null;
    phone: string | null;
    website_url: string | null;
    state: string;
    street: string | null;
  }
  